Incremental separation prep. Tests to be done with MIFNO at this point.
